style: license folders and clients renamed
Licenses -> licenses, clients -> managers, managers/xManager ->
managers/x
