--- clients/index.js
@@ -1,11 +1,0 @@
-const { CargoManager } = require('./cargoManager');
-const {NodeManager}= require('./nodeManager');
-const {NugetManager} = require('./nugetManager');
-const {PipManager} = require('./pipManager');
-
-module.exports = {
-    NodeManager,
-    NugetManager,
-    CargoManager,
-    PipManager,
-}
